Updated to latest TL source, updated to use Lib instead of copy
